[수정] storybook 폴더 안에 불필요하게 중복된 파일들 정리
  - components 폴더와 storybook 폴더 양쪽 모두에 파일이 있었음
  - storybook 폴더의 각 컴포넌트 파일들을 삭제하고 import 참조 위치를 변경하여, 중복을 제거함
