--- src/stories/components/atoms/index.js
@@ -1,6 +1,0 @@
-import TitleInput from './TitleInput';
-import TextColorButton from './TextColorButton';
-import TextEditButton from './TextEditButton';
-import TextBox from './TextBox';
-
-export { TitleInput, TextColorButton, TextEditButton, TextBox };
