--- src/stories/components/molecules/index.js
@@ -1,4 +1,0 @@
-import EditBox from './EditBox';
-import ColorBox from './ColorBox';
-
-export { EditBox, ColorBox };
