--- src/stories/components/organisms/index.js
@@ -1,3 +1,0 @@
-import TitleBox from './TitleBox';
-
-export { TitleBox };
