--- src/stories/components/pages/MainPage.js
@@ -1,27 +1,0 @@
-import styled from 'styled-components';
-import { TitleArea, ContentsArea } from '../templates';
-
-export default function MainPage() {
-  return (
-    <MainPageWrap className='MainPage'>
-      <TitleArea />
-      <ContentsArea />
-    </MainPageWrap>
-  );
-}
-
-// 여기서 issue 버튼에 대응하는 함수 등이 작성되어야 한다.
-// 그러므로 상태관리 역시 여기서 다 진행한다
-
-const MainPageWrap = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  flex-grow: 1;
-  overflow: auto;
-  margin: 0;
-  border: 1px solid black;
-  height: 100vh;
-
-  /* box-sizing: border-box; */
-`;
