--- src/stories/components/pages/index.js
@@ -1,3 +1,0 @@
-import MainPage from './MainPage';
-
-export { MainPage };
