--- src/stories/components/templates/index.js
@@ -1,4 +1,0 @@
-import TitleArea from './TitleArea';
-import ContentsArea from './ContentsArea';
-
-export { TitleArea, ContentsArea };
